Mesh_3: fix conversion warnings

+++ Mesh_3/include/CGAL/Mesh_3/Mesh_global_optimizer.h
@@ -733,7 +733,7 @@ operator()(int nb_iterations, Visitor visitor)
     //Pb with Freeze : sometimes a few vertices continue moving indefinitely
     //if the nb of moving vertices is < 1% of total nb AND does not decrease
     if(do_freeze_
-      && nb_vertices_moved < 0.005 * double(initial_vertices_nb)
+      && double(nb_vertices_moved) < 0.005 * double(initial_vertices_nb)
       && nb_vertices_moved == moving_vertices.size())
     {
       // we should stop because we are
